add IR trait, provide it to Parse as a generic param
The IR trait represents all possible productions as associated types.
It exists to make the production rules pluggable -- for example, the
IR generator may opt to represent Names as interned strings represented
by usize values, or as full in-place strings, or as boxed u8 arrays.

This also allows for _validating_ IR generators -- ones that keep an idea
of type/code state as they parse.

And also allows for IR generators that rewrite incoming code -- adding NOPs or
representing code as handles to be compiled by the an engine.